refactor: replace converter for course and form-data
